feat: Image default device is now cuda, if cuda is available; else it uses cpu docs: Added docstrings to all Image methods feat: Added IllegalFormatError refactor: Removed ImagePil
