+++ src/safeds/data/image/containers/__init__.py
@@ -1,9 +1,7 @@
 """Classes that can store image data."""
 
-from ._image_pil import ImagePil
 from ._image import Image
 
 __all__ = [
-    "ImagePil",
     "Image",
 ]
