+++ src/safeds/exceptions/_data.py
@@ -156,3 +156,10 @@ class ColumnIsTargetError(IllegalSchemaModificationError):
 
     def __init__(self, column_name: str) -> None:
         super().__init__(f'Column "{column_name}" is the target column and cannot be removed.')
+
+
+class IllegalFormatError(Exception):
+    """Exception raised when a format is not legal"""
+
+    def __init__(self, formats: list[str] | str, reason: str = "") -> None:
+        super().__init__(f"This format is illegal. {reason}Use one of the following formats: {formats}")
